+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- ref: Use a Context type mapping to a map[string]interface{} for all contexts
+
 ## v0.13.0
 
 - ref: Change DSN ProjectID to be a string (#420)